fix bug in BinaryDictionary.java which lead to crashes with dictionary using on 64bit devices; enable multi-dex and increase heap size also for Release build

+++ src/java/KP2ASoftkeyboard_AS/app/src/main/java/keepass2android/softkeyboard/BinaryDictionary.java
@@ -116,7 +116,7 @@ public class BinaryDictionary extends Dictionary {
         mDicTypeId = dicTypeId;
     }
 
-    private native int openNative(ByteBuffer bb, int typedLetterMultiplier,
+    private native long openNative(ByteBuffer bb, int typedLetterMultiplier,
                                   int fullWordMultiplier, int size);
